+++ src/image/oci/storage/image_store/image_store.h
@@ -98,7 +98,7 @@ int image_store_set_image_size(const char *id, uint64_t size);
 int image_store_get_all_images(imagetool_images_list *images_list);
 
 // Return the number of the known images.
-int image_store_get_images_number();
+size_t image_store_get_images_number();
 
 // Return a slice enumerating the images which have a big data
 // item with the name ImageDigestBigDataKey and the specified digest.
